+++ config/terraform/aws/ecs.tf
@@ -42,8 +42,10 @@ data "template_file" "covidportal_task" {
     invitation_email_template_id_fr        = aws_secretsmanager_secret_version.invitation_email_template_id_fr.arn
     password_reset_email_template_id_en    = aws_secretsmanager_secret_version.password_reset_email_template_id_en.arn
     password_reset_email_template_id_fr    = aws_secretsmanager_secret_version.password_reset_email_template_id_fr.arn
-    backup_code_admin_email_template_id_en = aws_secretsmanager_secret_version.backup_code_admin_email_template_id_en.arn,
-    backup_code_admin_email_template_id_fr = aws_secretsmanager_secret_version.backup_code_admin_email_template_id_fr.arn,
+    confirmation_email_template_id_en      = aws_secretsmanager_secret_version.confirmation_email_template_id_en.arn
+    confirmation_email_template_id_fr      = aws_secretsmanager_secret_version.confirmation_email_template_id_fr.arn
+    backup_code_admin_email_template_id_en = aws_secretsmanager_secret_version.backup_code_admin_email_template_id_en.arn
+    backup_code_admin_email_template_id_fr = aws_secretsmanager_secret_version.backup_code_admin_email_template_id_fr.arn
     django_secret_key                      = aws_secretsmanager_secret_version.django_secret_key.arn
     freshdesk_api_endpoint                 = aws_secretsmanager_secret_version.freshdesk_api_endpoint.arn
     freshdesk_api_key                      = aws_secretsmanager_secret_version.freshdesk_api_key.arn