+++ config/terraform/aws/secrets.tf
@@ -52,6 +52,26 @@ resource "aws_secretsmanager_secret_version" "invitation_email_template_id_fr" {
   secret_string = var.ecs_secret_invitation_email_template_id_fr
 }
 
+resource "aws_secretsmanager_secret" "confirmation_email_template_id_en" {
+  name                    = "confirmation_email_template_id_en"
+  recovery_window_in_days = 0
+}
+
+resource "aws_secretsmanager_secret_version" "confirmation_email_template_id_en" {
+  secret_id     = aws_secretsmanager_secret.confirmation_email_template_id_en.id
+  secret_string = var.ecs_secret_confirmation_email_template_id_en
+}
+
+resource "aws_secretsmanager_secret" "confirmation_email_template_id_fr" {
+  name                    = "confirmation_email_template_id_fr"
+  recovery_window_in_days = 0
+}
+
+resource "aws_secretsmanager_secret_version" "confirmation_email_template_id_fr" {
+  secret_id     = aws_secretsmanager_secret.confirmation_email_template_id_fr.id
+  secret_string = var.ecs_secret_confirmation_email_template_id_fr
+}
+
 resource "aws_secretsmanager_secret" "password_reset_email_template_id_en" {
   name                    = "password_reset_email_template_id_en"
   recovery_window_in_days = 0