+++ profiles/apps.py
@@ -1,4 +1,5 @@
 from django.apps import AppConfig
+from portal import container
 
 
 class ProfilesConfig(AppConfig):
@@ -6,3 +7,13 @@ class ProfilesConfig(AppConfig):
 
     def ready(self):
         import profiles.signals  # noqa
+
+        """
+        Wire the dependency-injector.
+        Each application must add this code in the apps.py file
+        and list each module that uses dependency-injections
+        """
+        from . import forms
+        from .utils import invitation_adapter
+
+        container.wire(modules=[forms, invitation_adapter])