+++ profiles/utils/invitation_adapter.py
@@ -2,8 +2,9 @@ from django.conf import settings
 from django.dispatch import Signal
 from invitations.adapters import BaseInvitationsAdapter
 
-from notifications_python_client.errors import HTTPError
-from notifications_python_client.notifications import NotificationsAPIClient
+from dependency_injector.wiring import inject, Provide
+from portal.containers import Container
+from portal.services import NotifyService
 
 user_signed_up = Signal(providing_args=["request", "user"])
 
@@ -12,21 +13,21 @@ class HealthcareInvitationAdapter(BaseInvitationsAdapter):
     def get_user_signed_up_signal(self):
         return user_signed_up
 
-    def send_mail(self, template_prefix, email, context):
-        notifications_client = NotificationsAPIClient(
-            settings.NOTIFY_API_KEY, base_url=settings.NOTIFY_ENDPOINT
+    @inject
+    def send_mail(
+        self,
+        template_prefix,
+        email,
+        context,
+        notify_service: NotifyService = Provide[Container.notify_service],
+    ):
+        notify_service.send_email(
+            address=context.get("email"),
+            template_id=settings.INVITATION_EMAIL_TEMPLATE_ID.get(
+                context.get("language") or "en"
+            ),
+            details={
+                "url": context.get("invite_url"),
+                "admin_email": context.get("inviter").email,
+            },
         )
-
-        try:
-            notifications_client.send_email_notification(
-                email_address=context.get("email"),
-                template_id=settings.INVITATION_EMAIL_TEMPLATE_ID.get(
-                    context.get("language") or "en"
-                ),
-                personalisation={
-                    "url": context.get("invite_url"),
-                    "admin_email": context.get("inviter").email,
-                },
-            )
-        except HTTPError as e:
-            raise Exception(e)