+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.15.0] - 2021-01-07
+
+### Added
+
+- Added a confirmation email that is sent to new users upon successful signup
+
+### Updated
+
+- Added package "dependency-injector" for managing the GC notification client as a service.
+  This is a code refactor and doesn't affect functionality.
+
 ## [1.14.0] - 2021-01-06
 
 ### Added